requirements.txt
aniso8601==9.0.1
APScheduler==3.7.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
DateTime==4.3
Flask==2.0.1
Flask-RESTful==0.3.9
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
zope.interface==5.4.0
